feat(#34): end-point 가구현
- 확인, 취소 대응을 위한 end-point 추가
